feat: :bug: Se corriguio el error
Hay que usar @JsonManagedReference Cundo es @OneToMany

+++ src/main/java/com/electro/electro_app/application/service/ICityService.java
@@ -8,11 +8,11 @@ import com.electro.electro_app.domain.entities.City;
 public interface ICityService {
     List<City> findAll();
 
-    Optional<City> findById(int id);
+    Optional<City> findById(Long id);
 
     City save(City city);
 
-    Optional<City> update(int id, City city);
+    Optional<City> update(Long id, City city);
 
-    Optional<City> delete(int id);
+    Optional<City> delete(Long id);
 }
