+++ src/main/java/com/electro/electro_app/infraestructure/repository/City/CityRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.electro.electro_app.domain.entities.City;
 
 @Repository
-public interface CityRepository extends JpaRepository<City, Integer> {
+public interface CityRepository extends JpaRepository<City, Long> {
     // Custom query methods can be defined here if needed
 
 }
